refactor: separated modules to individual files
