+++ app/src/main/java/ksc/campus/tech/kakao/map/domain/repositories/MapViewRepository.kt
@@ -1,14 +1,8 @@
 package ksc.campus.tech.kakao.map.domain.repositories
 
 import com.kakao.vectormap.camera.CameraPosition
-import dagger.Binds
-import dagger.Module
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
 import kotlinx.coroutines.flow.SharedFlow
-import ksc.campus.tech.kakao.map.data.repositoryimpls.MapViewRepositoryImpl
 import ksc.campus.tech.kakao.map.domain.models.LocationInfo
-import javax.inject.Singleton
 
 interface MapViewRepository {
     val selectedLocation: SharedFlow<LocationInfo?>
@@ -19,14 +13,4 @@ interface MapViewRepository {
     suspend fun updateCameraPositionWithFixedZoom(latitude: Double, longitude: Double)
     suspend fun updateCameraPosition(position: CameraPosition)
     suspend fun clearSelectedLocation()
-}
-
-@Module
-@InstallIn(SingletonComponent::class)
-abstract class MapRepositoryModule {
-    @Binds
-    @Singleton
-    abstract fun bindMapRepository(
-        mapRepositoryImpl: MapViewRepositoryImpl
-    ): MapViewRepository
 }
