+++ app/src/main/java/ksc/campus/tech/kakao/map/domain/repositories/SearchResultRepository.kt
@@ -1,27 +1,11 @@
 package ksc.campus.tech.kakao.map.domain.repositories
 
-import dagger.Binds
-import dagger.Module
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
 import kotlinx.coroutines.flow.SharedFlow
-import ksc.campus.tech.kakao.map.data.repositoryimpls.SearchResultRepositoryImpl
 import ksc.campus.tech.kakao.map.domain.models.SearchResult
-import javax.inject.Singleton
 
 
 interface SearchResultRepository {
     val searchResult: SharedFlow<List<SearchResult>>
 
     suspend fun search(text: String, apiKey: String)
-}
-
-@Module
-@InstallIn(SingletonComponent::class)
-abstract class SearchResultRepositoryModule {
-    @Binds
-    @Singleton
-    abstract fun bindSearchResultRepository(
-        searchResultRepositoryImpl: SearchResultRepositoryImpl
-    ): SearchResultRepository
 }
